+++ services/preview/visualstudio/mgmt/2014-04-01-preview/visualstudio/CHANGELOG.md
@@ -1,2 +1,22 @@
-# Change History
+# Unreleased Content
 
+## Breaking Changes
+
+### Signature Changes
+
+#### Struct Fields
+
+1. AccountResourceRequest.OperationType changed type from interface{} to AccountResourceRequestOperationType
+
+## Additive Changes
+
+### New Constants
+
+1. AccountResourceRequestOperationType.Create
+1. AccountResourceRequestOperationType.Link
+1. AccountResourceRequestOperationType.Unknown
+1. AccountResourceRequestOperationType.Update
+
+### New Funcs
+
+1. PossibleAccountResourceRequestOperationTypeValues() []AccountResourceRequestOperationType